# entity-containment/model_error_unknown_response.go
/*
 * Entity Containment API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package entitycontainment

// ErrorUnknownResponse - error returned for a 404 indicating one of the provided params is unknown
type ErrorUnknownResponse struct {

	// which subject is unknown
	Subject string `json:"subject"`

	// the value passed in that produced the unknown result
	Value string `json:"value"`

	// a human-readable string describing the failure
	Message string `json:"message"`
}

// AssertErrorUnknownResponseRequired checks if the required fields are not zero-ed
func AssertErrorUnknownResponseRequired(obj ErrorUnknownResponse) error {
	elements := map[string]interface{}{
		"subject": obj.Subject,
		"value":   obj.Value,
		"message": obj.Message,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertRecurseErrorUnknownResponseRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of ErrorUnknownResponse (e.g. [][]ErrorUnknownResponse), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseErrorUnknownResponseRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aErrorUnknownResponse, ok := obj.(ErrorUnknownResponse)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertErrorUnknownResponseRequired(aErrorUnknownResponse)
	})
}