# entity-containment/model_health_check_response.go
/*
 * Entity Containment API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package entitycontainment

type HealthCheckResponse struct {
	Status string `json:"status,omitempty"`

	Version string `json:"version,omitempty"`
}

// AssertHealthCheckResponseRequired checks if the required fields are not zero-ed
func AssertHealthCheckResponseRequired(obj HealthCheckResponse) error {
	return nil
}

// AssertRecurseHealthCheckResponseRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of HealthCheckResponse (e.g. [][]HealthCheckResponse), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseHealthCheckResponseRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aHealthCheckResponse, ok := obj.(HealthCheckResponse)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertHealthCheckResponseRequired(aHealthCheckResponse)
	})
}